requirements.txt
streamlit==1.3.0
pandas==1.3.5
matplotlib==3.5.1
plotly-express==0.4.1
